mem0-service/requirements.txt
mem0ai>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
psycopg2-binary>=2.9.9
redis>=5.0.0
aioredis>=2.0.0
